refactor(appBuilder): move WalkthroughNode
